+++ CHANGELOG.md
@@ -61,8 +61,11 @@ and this project adheres to
     in the racks (#72).
   - Introduce `DatacenterLocation` object attached to `Datacenter` optional
     location property to define GPS coordinates of the datacenters.
-- ui: Add map of datacenters if datacenters view with their geographical
-  positions if defined in database (#43).
+- ui:
+  - Add map of datacenters if datacenters view with their geographical
+    positions if defined in database (#43).
+  - Add possibility to filter equipment in infrastructure by rack, equipment
+    category, equipment type, equipment name and tags (#42).
 - pkg:
   - Add dependency on PyGObject python library (used to call Pango library).
   - Add optional dependency on requests-toolbelt python library on web variant